fixed/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
Sphinx==7.2.6
sphinx-rtd-theme==2.0.0
